Moved APICoverageReport, APICoverageRow, TestReport classes to separate files. Extracted the logic to generate the APICoverageReport to a separate class: ApiCoverageReportGenerator(for now). Added tests for ApiCoverageReportGenerator.
